deal with several signness issues
This minor commit contains 2 parts:
1. add -Wall compilation warnings to configure.ac
2. deal with a few signedness discrepancies (but probably no real world
security problems)

+++ src/crypto/CryptoTests.cpp
@@ -398,7 +398,7 @@ TEST_CASE("StrKey tests", "[crypto]")
         std::vector<uint8_t> in(input(expectedSize));
         std::string encoded = strKey::toStrKey(version, in);
 
-        for (int p = 0; p < encoded.size(); p++)
+        for (size_t p = 0u; p < encoded.size(); p++)
         {
             for (int st = 0; st < 4; st++)
             {
